// remotion-svg-primitives/src/lessons/legacy/generated/kp1FenYuHeIntroSilences.ts
export const kp1FenYuHeIntroSilences: SilenceInterval[] = [
  {
    "durationSeconds": 0.323,
    "endFrame": 30,
    "endSeconds": 0.985,
    "startFrame": 20,
    "startSeconds": 0.662
  },
  {
    "durationSeconds": 0.325,
    "endFrame": 89,
    "endSeconds": 2.973,
    "startFrame": 79,
    "startSeconds": 2.648
  },
  {
    "durationSeconds": 0.589,
    "endFrame": 130,
    "endSeconds": 4.333,
    "startFrame": 112,
    "startSeconds": 3.744
  },
  {
    "durationSeconds": 0.423,
    "endFrame": 154,
    "endSeconds": 5.121,
    "startFrame": 141,
    "startSeconds": 4.698
  },
  {
    "durationSeconds": 0.56,
    "endFrame": 232,
    "endSeconds": 7.736,
    "startFrame": 215,
    "startSeconds": 7.176
  },
  {
    "durationSeconds": 0.411,
    "endFrame": 297,
    "endSeconds": 9.884,
    "startFrame": 284,
    "startSeconds": 9.473
  },
  {
    "durationSeconds": 0.535,
    "endFrame": 345,
    "endSeconds": 11.501,
    "startFrame": 329,
    "startSeconds": 10.966
  },
  {
    "durationSeconds": 0.398,
    "endFrame": 398,
    "endSeconds": 13.265,
    "startFrame": 386,
    "startSeconds": 12.867
  },
  {
    "durationSeconds": 0.554,
    "endFrame": 469,
    "endSeconds": 15.636,
    "startFrame": 452,
    "startSeconds": 15.082
  },
  {
    "durationSeconds": 0.332,
    "endFrame": 521,
    "endSeconds": 17.354,
    "startFrame": 511,
    "startSeconds": 17.023
  },
  {
    "durationSeconds": 0.638,
    "endFrame": 577,
    "endSeconds": 19.222,
    "startFrame": 558,
    "startSeconds": 18.584
  },
  {
    "durationSeconds": 0.308,
    "endFrame": 598,
    "endSeconds": 19.919,
    "startFrame": 588,
    "startSeconds": 19.611
  },
  {
    "durationSeconds": 0.42,
    "endFrame": 633,
    "endSeconds": 21.087,
    "startFrame": 620,
    "startSeconds": 20.667
  },
  {
    "durationSeconds": 0.286,
    "endFrame": 672,
    "endSeconds": 22.396,
    "startFrame": 663,
    "startSeconds": 22.11
  },
  {
    "durationSeconds": 0.416,
    "endFrame": 723,
    "endSeconds": 24.091,
    "startFrame": 710,
    "startSeconds": 23.674
  },
  {
    "durationSeconds": 0.621,
    "endFrame": 810,
    "endSeconds": 27.004,
    "startFrame": 791,
    "startSeconds": 26.383
  },
  {
    "durationSeconds": 0.292,
    "endFrame": 864,
    "endSeconds": 28.786,
    "startFrame": 855,
    "startSeconds": 28.494
  },
  {
    "durationSeconds": 0.517,
    "endFrame": 915,
    "endSeconds": 30.494,
    "startFrame": 899,
    "startSeconds": 29.977
  },
  {
    "durationSeconds": 0.303,
    "endFrame": 965,
    "endSeconds": 32.157,
    "startFrame": 956,
    "startSeconds": 31.855
  },
  {
    "durationSeconds": 0.424,
    "endFrame": 1013,
    "endSeconds": 33.763,
    "startFrame": 1000,
    "startSeconds": 33.339
  },
  {
    "durationSeconds": 0.203,
    "endFrame": 1031,
    "endSeconds": 34.364,
    "startFrame": 1025,
    "startSeconds": 34.16
  },
  {
    "durationSeconds": 0.443,
    "endFrame": 1068,
    "endSeconds": 35.602,
    "startFrame": 1055,
    "startSeconds": 35.159
  },
  {
    "durationSeconds": 0.57,
    "endFrame": 1147,
    "endSeconds": 38.218,
    "startFrame": 1129,
    "startSeconds": 37.648
  },
  {
    "durationSeconds": 0.224,
    "endFrame": 1174,
    "endSeconds": 39.148,
    "startFrame": 1168,
    "startSeconds": 38.924
  },
  {
    "durationSeconds": 0.409,
    "endFrame": 1219,
    "endSeconds": 40.65,
    "startFrame": 1207,
    "startSeconds": 40.241
  },
  {
    "durationSeconds": 0.562,
    "endFrame": 1265,
    "endSeconds": 42.162,
    "startFrame": 1248,
    "startSeconds": 41.6
  },
  {
    "durationSeconds": 0.176,
    "endFrame": 1288,
    "endSeconds": 42.929,
    "startFrame": 1283,
    "startSeconds": 42.753
  },
  {
    "durationSeconds": 0.552,
    "endFrame": 1324,
    "endSeconds": 44.149,
    "startFrame": 1308,
    "startSeconds": 43.597
  },
  {
    "durationSeconds": 0.403,
    "endFrame": 1406,
    "endSeconds": 46.877,
    "startFrame": 1394,
    "startSeconds": 46.473
  },
  {
    "durationSeconds": 0.302,
    "endFrame": 1449,
    "endSeconds": 48.291,
    "startFrame": 1440,
    "startSeconds": 47.988
  }
];